+++ circuits/src/main/java/com/sk89q/craftbook/gates/logic/NotLowDelayer.java
@@ -37,7 +37,7 @@ public class NotLowDelayer extends AbstractIC {
     @Override
     public void trigger(final ChipState chip) {
 
-        int delay = Integer.parseInt(getSign().getLine(2));
+        long delay = Long.parseLong(getSign().getLine(2));
         if (chip.getInput(0)) {
             chip.setOutput(0, false);
         } else {
